pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Tony Okeke <tonykabilanokeke@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.7.2"
python-telegram-bot = "^20.7"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
asyncio = "^3.4.3"
rich = "^13.7.0"
watchdog = "^3.0.0"
aiohttp = "^3.9.3"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
